+++ app/src/main/cpp/FFmpegWrapper.h
@@ -33,7 +33,7 @@ public:
 
     audioParam getAPara();
     long int getDuration();
-    long int getCurrentPosition();
+    double getCurrentPosition();
 
     int FFmpegInitResample();
